requirements.txt
attrs==25.3.0
certifi==2025.7.14
dateparser==1.2.2
dnspython==2.7.0
exceptiongroup==1.3.0
h11==0.16.0
idna==3.10
ndjson==0.3.1
outcome==1.3.0.post0
pymongo==4.13.2
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2024.11.6
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
